Coordinator dart file added
In the end, I could not breach the gap between dart and kotlin. My class is the one that relied the most on the other ones, but the difference between the various languages was too much to compensate for. I made a skeleton version of the user class in dart so that my code can at least function.